Put CTestConfiguration type back in ctest_lab.ts
